+++ app/views/layouts/application.html.erb
@@ -4,7 +4,7 @@
     <title>Final</title>
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
-
+    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no">
     <%= stylesheet_link_tag    'application', media: 'all' %>
   </head>
