+++ client/src/assets/styles/global.scss
@@ -1,0 +1,18 @@
+$white: #fff;
+
+$blue:#62b5e5;
+$blue-regular: #68c4ed;
+$blue-light: #8ed3f1;
+$blue-lighter: #b3e1f6;
+
+$gray:#67899c;
+$gray-regular: #74879b;
+$gray-light: #000;
+$gray-lighter: #8090a0;
+
+$green:#addc91;
+$green-regular: #ace9a9;
+$green-light: #c1efbe;
+$green-lighter: #d5f4d4;
+
+$font: 'Helvalica Neue Light'
