+++ client/src/assets/styles/header.scss
@@ -1,0 +1,16 @@
+@import './global.scss';
+
+header {
+  display: flex;
+  justify-content: center;
+
+  h1 {
+    align-self: center;
+    text-align: center;
+    font-weight: normal;
+    color: $white;
+    margin: 0;
+    padding: 0;
+    font-size: 150%;
+  }
+}
