+++ client/src/components/Header.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+// Styles
+import '../assets/styles/header.scss';
+
+export default function Header() {
+  return (
+    <header>
+      <h1>DLTrackTask</h1>
+    </header>
+  );
+}
