+++ client/src/container/index.jsx
@@ -1,3 +1,3 @@
-import Header from './Header';
+import Footer from './Footer';
 
-export { Header };
+export { Footer };
