+++ client/src/pages/NotFound.jsx
@@ -4,11 +4,11 @@ import { LinkButton } from '../components';
 
 function NotFound() {
   return (
-    <div className="App">
+    <main>
       <h1>Error: 401</h1>
       <h2>Page Not Found</h2>
       <LinkButton to="/">Go Back Home</LinkButton>
-    </div>
+    </main>
   );
 }
